automation_app/requirements.txt
Appium-Python-Client==3.1.1
behave==1.2.6
selenium==4.18.1
pytest==8.0.0
python-dotenv==1.0.1
requests==2.31.0
